+++ src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { CalcController } from './calc/calculator.controller';
-
+import { CalcController } from './controllers/calculator.controller';
+import { ApiCalcController } from './modules/calculator/calculator.api.service';
 @Module({
   imports: [],
   controllers: [AppController, CalcController],
-  providers: [AppService],
+  providers: [AppService, ApiCalcController],
 })
 export class AppModule {}
