+++ src/controllers/dto/calculator.dto.ts
@@ -1,0 +1,4 @@
+export type calcData = {
+  status: string;
+  data: number;
+};
